use integer scaling for graphics

+++ counter.h
@@ -4,7 +4,7 @@
 #include <SDL.h>
 
 int draw_counter(SDL_Renderer *rend,
-		SDL_Point screen_pos, double size, SDL_Color color,
+		SDL_Point screen_pos, int size, SDL_Color color,
 		int number, int digits);
 
 #endif
